# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ZilliPay Currency Conversion Service",
        "description": "Currency exchanges by providing real-time conversion rates. It integrates two currency providers, CurrencyApi and CurrencyExchangeApi",
        "version": "1.0.0"
    },
    "paths": {
        "/api/currencyconvert/{to}/{from}/{amount}": {
            "get": {
                "tags": [
                    "Currency Conversion"
                ],
                "summary": "Convert currency from one type to another",
                "operationId": "c4f178e335b75160dc6289a6a6ac65e9",
                "parameters": [
                    {
                        "name": "to",
                        "in": "path",
                        "description": "Currency code to convert to (e.g. USD)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "USD"
                        }
                    },
                    {
                        "name": "from",
                        "in": "path",
                        "description": "Currency code to convert from (e.g. EUR)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "EUR"
                        }
                    },
                    {
                        "name": "amount",
                        "in": "path",
                        "description": "Amount to be converted",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful currency conversion",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "result": {
                                            "type": "number",
                                            "example": 84.75
                                        }
                                    },
                                    "type": "object"
                                },
                                "examples": {
                                    "result": {
                                        "summary": "Conversion result.",
                                        "value": {
                                            "success": true,
                                            "result": 84.75
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request or validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid currency code"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Currency Conversion",
            "description": "Currency Conversion"
        }
    ]
}